operator: Move SymbolRegistry from devito/core to devito/operator
