+++ devito/operator/__init__.py
@@ -1,3 +1,4 @@
+from .symbols import SymbolRegistry  # noqa
 from .operator import Operator  # noqa
 from .profiling import profiler_registry  # noqa
 from .registry import operator_registry  # noqa
